requirements.txt
apache-airflow
mysql-connector-python
cassandra-driver